+++ R/zzz.R
@@ -13,7 +13,7 @@
   .pkgenv$use_cube_cache = TRUE
   .pkgenv$threads = 1
   .pkgenv$debug = FALSE
-  .pkgenv$swarm = NULL
+  #.pkgenv$swarm = NULL
   
   # for windows, rwinlib includes GDAL data and PROJ data in the package and we must set the environment variables
   # PROJ_LIB and GDAL_DATA to make sure GDAL finds the data at the package location