+++ src/gdalcubes.cpp
@@ -126,13 +126,13 @@ struct error_handling_r {
     std::string code = (error_code != 0) ? " (" + std::to_string(error_code) + ")" : "";
     std::string where_str = (where.empty()) ? "" : " [in " + where + "]";
     if (type == error_level::ERRLVL_ERROR || type == error_level::ERRLVL_FATAL ) {
-      _err_stream << "Error message:\n"  << msg << where_str << std::endl;
+      _err_stream << "Error  message: "  << msg << where_str;
     } else if (type == error_level::ERRLVL_WARNING) {
-      _err_stream << "Warning message:\n" << msg << where_str << std::endl;
+      _err_stream << "Warning  message: " << msg << where_str;
     } else if (type == error_level::ERRLVL_INFO) {
-      _err_stream << "Info: " << msg << where_str << std::endl;
+      _err_stream << "Info message: " << msg << where_str;
     } else if (type == error_level::ERRLVL_DEBUG) {
-      _err_stream << "Debug message: "  << msg << where_str << std::endl;
+      _err_stream << "Debug message: "  << msg << where_str;
     }
     if (!_defer) {
       Rcpp::Rcerr << _err_stream.str() << std::endl;
@@ -144,13 +144,12 @@ struct error_handling_r {
   static void standard(error_level type, std::string msg, std::string where, int error_code) {
     _m_errhandl.lock();
     std::string code = (error_code != 0) ? " (" + std::to_string(error_code) + ")" : "";
-    std::string where_str = (where.empty()) ? "" : " [in " + where + "]";
     if (type == error_level::ERRLVL_ERROR || type == error_level::ERRLVL_FATAL) {
-      _err_stream << "Error message:\n" << msg << std::endl;
+      _err_stream << "Error: " << msg;
     } else if (type == error_level::ERRLVL_WARNING) {
-      _err_stream << "Warning message:\n" << msg  << std::endl;
+      _err_stream << "Warning: " << msg;
     } else if (type == error_level::ERRLVL_INFO) {
-      _err_stream << "Info: " <<  msg << std::endl;
+      _err_stream << "## " << msg;
     }
     if (!_defer) {
       Rcpp::Rcerr << _err_stream.str() << std::endl;
@@ -294,40 +293,40 @@ Rcpp::List libgdalcubes_cube_info( SEXP pin) {
     
     std::shared_ptr<cube> x = *aa;
     
-    Rcpp::CharacterVector d_name(3);
-    Rcpp::NumericVector d_low(3);
-    Rcpp::NumericVector d_high(3);
-    Rcpp::IntegerVector d_n(3);
-    Rcpp::IntegerVector d_chunk(3);
     
-    d_name[0] = "t";
-    d_low[0] = x->st_reference()->t0().to_double();
-    d_high[0] = x->st_reference()->t1().to_double();
-    d_n[0] = x->st_reference()->nt();
-    d_chunk[0] = x->chunk_size()[0];
+    Rcpp::List dims =
+      Rcpp::List::create(
+        Rcpp::Named("t")=
+          Rcpp::List::create(
+            Rcpp::Named("low") = x->st_reference()->t0().to_string(),
+            Rcpp::Named("high") = x->st_reference()->t1().to_string(),
+            Rcpp::Named("count") = x->st_reference()->nt(),
+            Rcpp::Named("pixel_size") = x->st_reference()->dt().to_string(),
+            Rcpp::Named("chunk_size") = x->chunk_size()[0]),
+        Rcpp::Named("y") =  
+          Rcpp::List::create(
+            Rcpp::Named("low") = x->st_reference()->bottom(),
+            Rcpp::Named("high") = x->st_reference()->top(),
+            Rcpp::Named("count") = x->st_reference()->ny(),
+            Rcpp::Named("pixel_size") = x->st_reference()->dy(),
+            Rcpp::Named("chunk_size") = x->chunk_size()[1]),
+        Rcpp::Named("x")=
+          Rcpp::List::create(
+            Rcpp::Named("low") = x->st_reference()->left(),
+            Rcpp::Named("high") = x->st_reference()->right(),
+            Rcpp::Named("count") = x->st_reference()->nx(),
+            Rcpp::Named("pixel_size") = x->st_reference()->dx(),
+            Rcpp::Named("chunk_size") = x->chunk_size()[2])
+      );
+                         
+                         
+                      
+                       
     
-    d_name[1] = "y";
-    d_low[1] = x->st_reference()->bottom();
-    d_high[1] = x->st_reference()->top();
-    d_n[1] = x->st_reference()->ny();
-    d_chunk[1] = x->chunk_size()[1];
     
-    d_name[2] = "x";
-    d_low[2] = x->st_reference()->left();
-    d_high[2] = x->st_reference()->right();
-    d_n[2] = x->st_reference()->nx();
-    d_chunk[2] = x->chunk_size()[2];
-    
-    
-    Rcpp::DataFrame dims =
-      Rcpp::DataFrame::create(Rcpp::Named("name")=d_name,
-                              Rcpp::Named("low")=d_low,
-                              Rcpp::Named("high")=d_high,
-                              Rcpp::Named("size")=d_n,
-                              Rcpp::Named("chunk_size")=d_chunk);
     
     Rcpp::CharacterVector b_names(x->bands().count(), "");
-    Rcpp::CharacterVector b_type(x->bands().count(), "");
+    //Rcpp::CharacterVector b_type(x->bands().count(), "");
     Rcpp::NumericVector b_offset(x->bands().count(), NA_REAL);
     Rcpp::NumericVector b_scale(x->bands().count(), NA_REAL);
     Rcpp::NumericVector b_nodata(x->bands().count(), NA_REAL);
@@ -335,7 +334,7 @@ Rcpp::List libgdalcubes_cube_info( SEXP pin) {
     
     for (uint16_t i=0; i<x->bands().count(); ++i) {
       b_names[i] = x->bands().get(i).name;
-      b_type[i] = x->bands().get(i).type;
+     // b_type[i] = x->bands().get(i).type;
       b_offset[i] = x->bands().get(i).offset;
       b_scale[i] = x->bands().get(i).scale;
       b_nodata[i] = (x->bands().get(i).no_data_value.empty())? NA_REAL : std::stod(x->bands().get(i).no_data_value);
@@ -344,7 +343,7 @@ Rcpp::List libgdalcubes_cube_info( SEXP pin) {
     
     Rcpp::DataFrame bands =
       Rcpp::DataFrame::create(Rcpp::Named("name")=b_names,
-                              Rcpp::Named("type")=b_type,
+                              //Rcpp::Named("type")=b_type,
                               Rcpp::Named("offset")=b_offset,
                               Rcpp::Named("scale")=b_scale,
                               Rcpp::Named("nodata")=b_nodata,
@@ -363,7 +362,7 @@ Rcpp::List libgdalcubes_cube_info( SEXP pin) {
                               Rcpp::Named("srs") = x->st_reference()->srs(),
                               Rcpp::Named("proj4") = sproj4,
                               Rcpp::Named("graph") = x->make_constructible_json().dump(2),
-                              Rcpp::Named("size") = Rcpp::IntegerVector::create(x->size()[0], x->size()[1], x->size()[2], x->size()[3]));
+                              Rcpp::Named("size") = Rcpp::IntegerVector::create(x->size()[0], x->size()[1], x->size()[2], x->size()[3])); // TODO: remove size element
     
   }
   catch (std::string s) {
@@ -619,7 +618,7 @@ Rcpp::List libgdalcubes_image_collection_info( SEXP pin) {
     
     
     
-    std::vector<image_collection::bands_row> bands = ic->get_bands();
+    std::vector<image_collection::bands_row> bands = ic->get_available_bands();
     
     
     Rcpp::CharacterVector bands_name( bands.size());
@@ -1209,11 +1208,11 @@ SEXP libgdalcubes_create_select_bands_cube(SEXP pin, std::vector<std::string> ba
 }
 
 // [[Rcpp::export]]
-SEXP libgdalcubes_create_apply_pixel_cube(SEXP pin, std::vector<std::string> expr, std::vector<std::string> names) {
+SEXP libgdalcubes_create_apply_pixel_cube(SEXP pin, std::vector<std::string> expr, std::vector<std::string> names, bool keep_bands=false) {
   try {
     Rcpp::XPtr< std::shared_ptr<cube> > aa = Rcpp::as<Rcpp::XPtr<std::shared_ptr<cube>>>(pin);
     
-    std::shared_ptr<apply_pixel_cube>* x = new std::shared_ptr<apply_pixel_cube>(apply_pixel_cube::create(*aa, expr, names));
+    std::shared_ptr<apply_pixel_cube>* x = new std::shared_ptr<apply_pixel_cube>(apply_pixel_cube::create(*aa, expr, names, keep_bands));
     Rcpp::XPtr< std::shared_ptr<apply_pixel_cube> > p(x, true) ;
     
     return p;
@@ -1225,10 +1224,10 @@ SEXP libgdalcubes_create_apply_pixel_cube(SEXP pin, std::vector<std::string> exp
 }
 
 // [[Rcpp::export]]
-SEXP libgdalcubes_create_stream_apply_pixel_cube(SEXP pin, std::string cmd, uint16_t nbands, std::vector<std::string> names) {
+SEXP libgdalcubes_create_stream_apply_pixel_cube(SEXP pin, std::string cmd, uint16_t nbands, std::vector<std::string> names, bool keep_bands = false) {
   try {
     Rcpp::XPtr< std::shared_ptr<cube> > aa = Rcpp::as<Rcpp::XPtr<std::shared_ptr<cube>>>(pin);
-    std::shared_ptr<stream_apply_pixel_cube>* x = new std::shared_ptr<stream_apply_pixel_cube>(stream_apply_pixel_cube::create(*aa, cmd, nbands, names));
+    std::shared_ptr<stream_apply_pixel_cube>* x = new std::shared_ptr<stream_apply_pixel_cube>(stream_apply_pixel_cube::create(*aa, cmd, nbands, names, keep_bands));
     Rcpp::XPtr< std::shared_ptr<stream_apply_pixel_cube> > p(x, true) ;
     return p;
   }
@@ -1273,10 +1272,37 @@ void libgdalcubes_debug_output( bool debug) {
 
 
 // [[Rcpp::export]]
-void libgdalcubes_eval_cube( SEXP pin, std::string outfile, uint8_t compression_level=0) {
+void libgdalcubes_eval_cube( SEXP pin, std::string outfile, uint8_t compression_level=0, bool with_VRT=false) {
   try {
     Rcpp::XPtr< std::shared_ptr<cube> > aa = Rcpp::as<Rcpp::XPtr< std::shared_ptr<cube> >>(pin);
-    (*aa)->write_netcdf_file(outfile, compression_level);
+    (*aa)->write_netcdf_file(outfile, compression_level, with_VRT);
+  }
+  catch (std::string s) {
+    Rcpp::stop(s);
+  }
+}
+
+// [[Rcpp::export]]
+void libgdalcubes_write_COG( SEXP pin, std::string dir, std::string prefix="", std::string rsmpl_overview = "nearest", SEXP creation_options = R_NilValue ) {
+  try {
+    Rcpp::XPtr< std::shared_ptr<cube> > aa = Rcpp::as<Rcpp::XPtr< std::shared_ptr<cube> >>(pin);
+    std::map<std::string, std::string> co;
+    
+    if (creation_options == R_NilValue) {
+      (*aa)->write_COG_collection(dir, prefix, rsmpl_overview);
+    }
+    else {
+      Rcpp::List colist = Rcpp::as<Rcpp::List>(creation_options);
+      Rcpp::CharacterVector names = colist.names();
+      for(uint16_t i=0; i< names.size(); ++i) {
+        std::string key = Rcpp::as<std::string>(names[i]);
+        std::string value =Rcpp::as<std::string>(Rcpp::as<Rcpp::CharacterVector>(colist[i]));
+        co[key] = value;
+      }
+      (*aa)->write_COG_collection(dir, prefix, rsmpl_overview, co);
+    }
+    
+    
   }
   catch (std::string s) {
     Rcpp::stop(s);
@@ -1324,6 +1350,11 @@ void libgdalcubes_set_swarm(std::vector<std::string> swarm) {
   //p->set_threads(nthreads);
   config::instance()->set_default_chunk_processor(p);
   // TODO: how to make swarm interruptible
+}
+
+// [[Rcpp::export]]
+std::string libgdalcubes_simple_hash(std::string instr) {
+  return utils::hash(instr);
 }