+++ NEWS.md
@@ -1,11 +1,21 @@
 # gdalcubes 0.1.9999
 
 * rename `filter_predicate()` -> `filter_pixel()`
-* new operator `fill_time()` to fill NA pixels of data cubes based on time series interpolation
+* new operator `fill_time()` fills NA pixels of data cubes based on time series interpolation
 * add masking based on pixel band values while reading images, see `?image_mask`
 * collection format Sentinel2_L2A now includes WVP, AOT, and SCL bands 
 * add `write_json_descr`argument to `write_ncdf()`
-* `write_ncdf`, `as_stars()`, and `plot()` are now  interruptible, though it might take some time to let all threads finish their current chunk
+* new argument `with_VRT` in `write_ncdf()` to write GDAL VRT datasets for data cube time slices
+* `write_ncdf`, `as_stars()`, and `plot()` are now  interruptible, though it might take some time to let all threads finish processing their current chunk
+* new image collection database schema, existing collections must be recreated
+* add `animate()` function to create data cube time series animations
+* fix windows source compilation on CRAN
+* consistent output for printing data cube views and data cubes
+* new collection format for Sentinel-2 data on Theia
+* add `as_array()` function to convert a data cube to a native in-memory R array
+* collection formats can now overwrite scale, offset, and unit for bands
+* add `write_COG` to export data cube time slices as cloud-optimized GeoTIFF files
+
 
 
 # gdalcubes 0.1.0 (2019-05-15)